webkit master

+++ Source/WebKit/NetworkProcess/cache/CacheStorageEngineCaches.cpp
@@ -50,17 +50,22 @@ static inline String cachesOriginFilename(const String& cachesRootPath)
     return WebCore::FileSystem::pathByAppendingComponent(cachesRootPath, ASCIILiteral("origin"));
 }
 
-void Caches::retrieveOriginFromDirectory(const String& folderPath, WorkQueue& queue, WTF::CompletionHandler<void(std::optional<WebCore::SecurityOriginData>&&)>&& completionHandler)
+Caches::~Caches()
 {
-    queue.dispatch([completionHandler = WTFMove(completionHandler), folderPath = folderPath.isolatedCopy()]() mutable {
-        if (!WebCore::FileSystem::fileExists(cachesListFilename(folderPath))) {
+    ASSERT(m_pendingWritingCachesToDiskCallbacks.isEmpty());
+}
+
+void Caches::retrieveOriginFromDirectory(const String& folderPath, WorkQueue& queue, WTF::CompletionHandler<void(std::optional<WebCore::ClientOrigin>&&)>&& completionHandler)
+{
+    queue.dispatch([completionHandler = WTFMove(completionHandler), filename = cachesOriginFilename(folderPath)]() mutable {
+        if (!WebCore::FileSystem::fileExists(filename)) {
             RunLoop::main().dispatch([completionHandler = WTFMove(completionHandler)]() mutable {
                 completionHandler(std::nullopt);
             });
             return;
         }
 
-        auto channel = IOChannel::open(cachesOriginFilename(folderPath), IOChannel::Type::Read);
+        auto channel = IOChannel::open(filename, IOChannel::Type::Read);
         channel->read(0, std::numeric_limits<size_t>::max(), nullptr, [completionHandler = WTFMove(completionHandler)](const Data& data, int error) mutable {
             ASSERT(RunLoop::isMain());
             if (error) {
@@ -72,9 +77,9 @@ void Caches::retrieveOriginFromDirectory(const String& folderPath, WorkQueue& qu
     });
 }
 
-Caches::Caches(Engine& engine, String&& origin, String&& rootPath, uint64_t quota)
+Caches::Caches(Engine& engine, WebCore::ClientOrigin&& origin, String&& rootPath, uint64_t quota)
     : m_engine(&engine)
-    , m_origin(WebCore::SecurityOriginData::fromSecurityOrigin(WebCore::SecurityOrigin::createFromString(origin)))
+    , m_origin(WTFMove(origin))
     , m_rootPath(WTFMove(rootPath))
     , m_quota(quota)
 {
@@ -83,27 +88,36 @@ Caches::Caches(Engine& engine, String&& origin, String&& rootPath, uint64_t quot
 void Caches::storeOrigin(CompletionCallback&& completionHandler)
 {
     WTF::Persistence::Encoder encoder;
-    encoder << m_origin.protocol;
-    encoder << m_origin.host;
-    encoder << m_origin.port;
+    encoder << m_origin.topOrigin.protocol;
+    encoder << m_origin.topOrigin.host;
+    encoder << m_origin.topOrigin.port;
+    encoder << m_origin.clientOrigin.protocol;
+    encoder << m_origin.clientOrigin.host;
+    encoder << m_origin.clientOrigin.port;
     m_engine->writeFile(cachesOriginFilename(m_rootPath), Data { encoder.buffer(), encoder.bufferSize() }, [protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)] (std::optional<Error>&& error) mutable {
         completionHandler(WTFMove(error));
     });
 }
 
-std::optional<WebCore::SecurityOriginData> Caches::readOrigin(const Data& data)
+std::optional<WebCore::ClientOrigin> Caches::readOrigin(const Data& data)
 {
     // FIXME: We should be able to use modern decoders for persistent data.
-    WebCore::SecurityOriginData origin;
+    WebCore::SecurityOriginData topOrigin, clientOrigin;
     WTF::Persistence::Decoder decoder(data.data(), data.size());
 
-    if (!decoder.decode(origin.protocol))
+    if (!decoder.decode(topOrigin.protocol))
         return std::nullopt;
-    if (!decoder.decode(origin.host))
+    if (!decoder.decode(topOrigin.host))
         return std::nullopt;
-    if (!decoder.decode(origin.port))
+    if (!decoder.decode(topOrigin.port))
         return std::nullopt;
-    return WTFMove(origin);
+    if (!decoder.decode(clientOrigin.protocol))
+        return std::nullopt;
+    if (!decoder.decode(clientOrigin.host))
+        return std::nullopt;
+    if (!decoder.decode(clientOrigin.port))
+        return std::nullopt;
+    return WebCore::ClientOrigin { WTFMove(topOrigin), WTFMove(clientOrigin) };
 }
 
 void Caches::initialize(WebCore::DOMCacheEngine::CompletionCallback&& callback)
@@ -130,47 +144,58 @@ void Caches::initialize(WebCore::DOMCacheEngine::CompletionCallback&& callback)
         callback(Error::WriteDisk);
         return;
     }
+
+    m_pendingInitializationCallbacks.append(WTFMove(callback));
     m_storage = storage.releaseNonNull();
     m_storage->writeWithoutWaiting();
 
-    storeOrigin([this, callback = WTFMove(callback)] (std::optional<Error>&& error) mutable {
+    storeOrigin([this] (std::optional<Error>&& error) mutable {
         if (error) {
-            callback(Error::WriteDisk);
+            auto pendingCallbacks = WTFMove(m_pendingInitializationCallbacks);
+            for (auto& callback : pendingCallbacks)
+                callback(Error::WriteDisk);
+
+            m_storage = nullptr;
             return;
         }
 
-        readCachesFromDisk([this, callback = WTFMove(callback)](Expected<Vector<Cache>, Error>&& result) mutable {
+        readCachesFromDisk([this](Expected<Vector<Cache>, Error>&& result) mutable {
             makeDirty();
 
             if (!result.has_value()) {
-                callback(result.error());
-
                 auto pendingCallbacks = WTFMove(m_pendingInitializationCallbacks);
                 for (auto& callback : pendingCallbacks)
                     callback(result.error());
+
+                m_storage = nullptr;
                 return;
             }
             m_caches = WTFMove(result.value());
 
-            initializeSize(WTFMove(callback));
+            initializeSize();
         });
     });
 }
 
-void Caches::initializeSize(WebCore::DOMCacheEngine::CompletionCallback&& callback)
+void Caches::initializeSize()
 {
     if (!m_storage) {
-        callback(Error::Internal);
+        auto pendingCallbacks = WTFMove(m_pendingInitializationCallbacks);
+        for (auto& callback : pendingCallbacks)
+            callback(Error::Internal);
         return;
     }
 
     uint64_t size = 0;
-    m_storage->traverse({ }, 0, [protectedThis = makeRef(*this), this, protectedStorage = makeRef(*m_storage), callback = WTFMove(callback), size](const auto* storage, const auto& information) mutable {
+    m_storage->traverse({ }, 0, [protectedThis = makeRef(*this), this, protectedStorage = makeRef(*m_storage), size](const auto* storage, const auto& information) mutable {
         if (!storage) {
+            if (m_pendingInitializationCallbacks.isEmpty()) {
+                // Caches was cleared so let's not get initialized.
+                m_storage = nullptr;
+                return;
+            }
             m_size = size;
             m_isInitialized = true;
-            callback(std::nullopt);
-
             auto pendingCallbacks = WTFMove(m_pendingInitializationCallbacks);
             for (auto& callback : pendingCallbacks)
                 callback(std::nullopt);
@@ -187,14 +212,26 @@ void Caches::detach()
 {
     m_engine = nullptr;
     m_rootPath = { };
+    clearPendingWritingCachesToDiskCallbacks();
 }
 
 void Caches::clear(CompletionHandler<void()>&& completionHandler)
 {
+    if (m_isWritingCachesToDisk) {
+        m_pendingWritingCachesToDiskCallbacks.append([this, completionHandler = WTFMove(completionHandler)] (auto&& error) mutable {
+            this->clear(WTFMove(completionHandler));
+        });
+        return;
+    }
+
+    auto pendingCallbacks = WTFMove(m_pendingInitializationCallbacks);
+    for (auto& callback : pendingCallbacks)
+        callback(Error::Internal);
+
     if (m_engine)
         m_engine->removeFile(cachesListFilename(m_rootPath));
     if (m_storage) {
-        m_storage->clear(String { }, std::chrono::system_clock::time_point::min(), [protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)]() mutable {
+        m_storage->clear(String { }, -WallTime::infinity(), [protectedThis = makeRef(*this), completionHandler = WTFMove(completionHandler)]() mutable {
             ASSERT(RunLoop::isMain());
             protectedThis->clearMemoryRepresentation();
             completionHandler();
@@ -202,7 +239,15 @@ void Caches::clear(CompletionHandler<void()>&& completionHandler)
         return;
     }
     clearMemoryRepresentation();
+    clearPendingWritingCachesToDiskCallbacks();
     completionHandler();
+}
+
+void Caches::clearPendingWritingCachesToDiskCallbacks()
+{
+    auto pendingWritingCachesToDiskCallbacks = WTFMove(m_pendingWritingCachesToDiskCallbacks);
+    for (auto& callback : pendingWritingCachesToDiskCallbacks)
+        callback(Error::Internal);
 }
 
 Cache* Caches::find(const String& name)
@@ -225,6 +270,17 @@ void Caches::open(const String& name, CacheIdentifierCallback&& callback)
 {
     ASSERT(m_isInitialized);
     ASSERT(m_engine);
+
+    if (m_isWritingCachesToDisk) {
+        m_pendingWritingCachesToDiskCallbacks.append([this, name, callback = WTFMove(callback)] (auto&& error) mutable {
+            if (error) {
+                callback(makeUnexpected(error.value()));
+                return;
+            }
+            this->open(name, WTFMove(callback));
+        });
+        return;
+    }
 
     if (auto* cache = find(name)) {
         cache->open([cacheIdentifier = cache->identifier(), callback = WTFMove(callback)](std::optional<Error>&& error) mutable {
@@ -252,11 +308,22 @@ void Caches::remove(uint64_t identifier, CacheIdentifierCallback&& callback)
     ASSERT(m_isInitialized);
     ASSERT(m_engine);
 
+    if (m_isWritingCachesToDisk) {
+        m_pendingWritingCachesToDiskCallbacks.append([this, identifier, callback = WTFMove(callback)] (auto&& error) mutable {
+            if (error) {
+                callback(makeUnexpected(error.value()));
+                return;
+            }
+            this->remove(identifier, WTFMove(callback));
+        });
+        return;
+    }
+
     auto position = m_caches.findMatching([&](const auto& item) { return item.identifier() == identifier; });
 
     if (position == notFound) {
         ASSERT(m_removedCaches.findMatching([&](const auto& item) { return item.identifier() == identifier; }) != notFound);
-        callback(CacheIdentifierOperationResult { identifier, false });
+        callback(CacheIdentifierOperationResult { 0, false });
         return;
     }
 
@@ -362,6 +429,8 @@ void Caches::readCachesFromDisk(WTF::Function<void(Expected<Vector<Cache>, Error
 
 void Caches::writeCachesToDisk(CompletionCallback&& callback)
 {
+    ASSERT(!m_isWritingCachesToDisk);
+    ASSERT(m_isInitialized);
     if (!shouldPersist()) {
         callback(std::nullopt);
         return;
@@ -375,8 +444,12 @@ void Caches::writeCachesToDisk(CompletionCallback&& callback)
         return;
     }
 
-    m_engine->writeFile(cachesListFilename(m_rootPath), encodeCacheNames(m_caches), [callback = WTFMove(callback)](std::optional<Error>&& error) mutable {
+    m_isWritingCachesToDisk = true;
+    m_engine->writeFile(cachesListFilename(m_rootPath), encodeCacheNames(m_caches), [this, protectedThis = makeRef(*this), callback = WTFMove(callback)](std::optional<Error>&& error) mutable {
+        m_isWritingCachesToDisk = false;
         callback(WTFMove(error));
+        while (!m_pendingWritingCachesToDiskCallbacks.isEmpty() && !m_isWritingCachesToDisk)
+            m_pendingWritingCachesToDiskCallbacks.takeFirst()(std::nullopt);
     });
 }
 
@@ -412,10 +485,11 @@ void Caches::writeRecord(const Cache& cache, const RecordInformation& recordInfo
 
     if (!shouldPersist()) {
         m_volatileStorage.set(recordInformation.key, WTFMove(record));
+        callback(std::nullopt);
         return;
     }
 
-    m_storage->store(Cache::encode(recordInformation, record), [protectedStorage = makeRef(*m_storage), callback = WTFMove(callback)](const Data&) {
+    m_storage->store(Cache::encode(recordInformation, record), { }, [protectedStorage = makeRef(*m_storage), callback = WTFMove(callback)]() {
         callback(std::nullopt);
     });
 }
@@ -473,10 +547,19 @@ void Caches::removeCacheEntry(const NetworkCache::Key& key)
 
 void Caches::clearMemoryRepresentation()
 {
+    if (!m_isInitialized) {
+        ASSERT(m_caches.isEmpty());
+        // m_storage might not be null in case Caches is being initialized. This is fine as nullify it below is a memory optimization.
+        return;
+    }
+
     makeDirty();
     m_caches.clear();
     m_isInitialized = false;
+
+    // Clear storages as a memory optimization.
     m_storage = nullptr;
+    m_volatileStorage.clear();
 }
 
 bool Caches::isDirty(uint64_t updateCounter) const
@@ -496,15 +579,26 @@ const NetworkCache::Salt& Caches::salt() const
     return m_volatileSalt.value();
 }
 
-CacheInfos Caches::cacheInfos(uint64_t updateCounter) const
+void Caches::cacheInfos(uint64_t updateCounter, CacheInfosCallback&& callback)
 {
+    if (m_isWritingCachesToDisk) {
+        m_pendingWritingCachesToDiskCallbacks.append([this, updateCounter, callback = WTFMove(callback)] (auto&& error) mutable {
+            if (error) {
+                callback(makeUnexpected(error.value()));
+                return;
+            }
+            this->cacheInfos(updateCounter, WTFMove(callback));
+        });
+        return;
+    }
+
     Vector<CacheInfo> cacheInfos;
     if (isDirty(updateCounter)) {
         cacheInfos.reserveInitialCapacity(m_caches.size());
         for (auto& cache : m_caches)
             cacheInfos.uncheckedAppend(CacheInfo { cache.identifier(), cache.name() });
     }
-    return { WTFMove(cacheInfos), m_updateCounter };
+    callback(CacheInfos { WTFMove(cacheInfos), m_updateCounter });
 }
 
 void Caches::appendRepresentation(StringBuilder& builder) const
